// src/features/loans/types.ts
import { z } from "zod";
import { loansSchema } from "./schema";

// z.input descreve o que o schema aceita na entrada (string)
export type LoansFormValues = z.input<typeof loansSchema>;

export type LoansSummary = {
    type: string;
    method: MethodType;
    amount: number;
    downPayment: number;
    extraAmortization?: number,
    extraAmortizationType?: ExtraAmortizationType,
    extraAmortizationMonth?: string
    avgInstallments: number;
    firstInstallment: number;
    annualRate: number;
    monthlyRate: number;
    iofCeiling: number;
    fixedIofPct: number,
    dailyIofPct: number,
    fixedIof: number,
    dailyIof: number,
    totalIof: number,
    iofWasCapped: boolean;
    totalPaidNoIof: number;
    totalPaidWithIof: number;
    totalInterest: number;
    totalInterestWithIof: number;
    startDate: string,
    insurancePercent: number,
    insuranceValue: number,
    cet: number,
};

export type ExtraAmortizationType = "reduzir_prazo" | "reduzir_parcela"
export type MethodType = "price" | "sac";

export interface LoansSchedule {
    payment: number
    balance: number
    principal: number
    interest?: number
    admin?: number
    quota?: number
    monthly?: number
    totalPaid?: number
    totalInterest?: number
    // eslint-disable-next-line @typescript-eslint/no-explicit-any
    [key: string]: any
}